pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "base58>=2.1.1",
    "httpx>=0.28.1",
    "pynacl>=1.6.1",
    "questionary>=2.1.1",
    "rich>=14.2.0",
    "solana>=0.36.10",
    "solders>=0.27.1",
]
